Optionally load properties from environment (#106)
Adds an optional argument `environmentPrefix` to `loadProperties`.
If this variable is non-null, environment variables are filtered
and parsed as properties, so the variable `PREFIX_APP_CFG` is loaded
as the property `app.cfg`. Setting the prefix to an empty string
will load all environment variables as properties.

The precedence of this loader is set fairly low, so environment
variables will override `application.properties`, but be overridden
by values from SSM Parameter Store or `overrides.properties`.

This change allows the user to pass configuration through normal
environment variables, which is an attractive option for containerized
services running in Docker or ECS.